requirements.txt
amqp==5.1.1
asgiref==3.7.2
asyncio==3.4.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
kombu==5.2.4
numpy==1.24.3
pandas==2.0.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
